241029 : 기존에 팝업 형태에서 마이페이지에서 수정 페이지로 이동 후 수정과 삭제로 변경 + 마이페이지에서 수정페이지 또는 새로운 글 작성 페이지로 페이지 이동할 때 스피너로 설정되어있는데, 이를 고정된 텍스트뷰로 바꿈 + 스피너와 검색의 영역이 겹치지 않도록 수정 + MyTextfileActivity의 파일이름을 MyNewTextActivity로 바꿈 + MyExistTextActivity.java 추가
