+++ app/src/main/res/layout/activity_my_new_text.xml
@@ -3,21 +3,21 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyTextfileActivity">
+    tools:context=".MyNewTextActivity">
 
-    <!-- 카테고리 선택 스피너 -->
-    <Spinner
-        android:id="@+id/categorySpinner"
-        android:layout_width="match_parent"
+    <!-- 카테고리 표시 텍스트뷰 -->
+    <TextView
+        android:id="@+id/categoryTextView"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:minHeight="48dp"
         android:padding="8dp"
+        android:textSize="16sp"
+        android:textColor="@color/black"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         android:layout_marginTop="50dp"
         android:layout_marginStart="16dp"
         android:layout_marginEnd="16dp" />
-
 
     <!-- 제목 입력 필드 -->
     <EditText
@@ -26,16 +26,13 @@
         android:layout_height="wrap_content"
         android:hint="제목을 입력하세요."
         android:layout_marginTop="16dp"
-        app:layout_constraintTop_toBottomOf="@id/categorySpinner"
+        app:layout_constraintTop_toBottomOf="@id/categoryTextView"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         android:padding="16dp"
         app:layout_constraintWidth_percent="0.9" />
 
     <!-- 글 내용 입력 필드 -->
-
-    <!-- 글 작성 버튼 -->
-
     <EditText
         android:id="@+id/postContentEditText"
         android:layout_width="0dp"
@@ -51,15 +48,16 @@
         app:layout_constraintTop_toBottomOf="@id/postTitleEditText"
         app:layout_constraintWidth_percent="0.9" />
 
+    <!-- 글 작성 버튼 -->
     <Button
         android:id="@+id/submitPostButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="글 작성"
+        android:backgroundTint="@color/colorAccent"
         app:layout_constraintTop_toBottomOf="@id/postContentEditText"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
         android:layout_margin="16dp" />
 
     <!-- 서식 버튼 추가 (볼드, 이탤릭) -->
@@ -86,4 +84,5 @@
             android:text="I"
             android:textStyle="italic" />
     </LinearLayout>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
