+++ app/src/main/res/layout/activity_my_page.xml
@@ -23,7 +23,7 @@
     <!-- 카테고리 스피너 -->
     <Spinner
         android:id="@+id/categorySpinner"
-        android:layout_width="wrap_content"
+        android:layout_width="150dp"
         android:layout_height="wrap_content"
         app:layout_constraintTop_toBottomOf="@id/myPageTitle"
         app:layout_constraintStart_toStartOf="parent"
@@ -72,7 +72,7 @@
         android:layout_height="wrap_content"
         android:text="글 작성"
         android:backgroundTint="@color/colorAccent"
-        android:textColor="#FFFFFF"
+        android:textColor="@android:color/black"
         android:textSize="16sp"
         android:layout_margin="16dp"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -112,7 +112,7 @@
         android:visibility="gone"
         android:padding="8dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.783"
+        app:layout_constraintHorizontal_bias="0.85"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/myPageTitle" />
